requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.3.42
openai==1.78.0
orjson==3.10.18
packaging==24.2
psutil==7.0.0
pydantic==2.11.4
pydantic-core==2.33.2
python-dotenv==1.1.0
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
zstandard==0.23.0
